Move PProfOStream from InstrProfWriter.cpp to InstrProf.h/cpp
